Move the checker package into the compiler package.
