+++ cmd/mdot/main.go
@@ -29,8 +29,8 @@ import (
 
 	"github.com/golang/glog"
 	"github.com/google/mtail/internal/mtail"
-	"github.com/google/mtail/internal/vm/checker"
 	"github.com/google/mtail/internal/vm/compiler/ast"
+	"github.com/google/mtail/internal/vm/compiler/checker"
 	"github.com/google/mtail/internal/vm/parser"
 )
