+++ docs/debugging.md
@@ -99,7 +99,7 @@ The formatted mtail program should help make it obvious what's happening and let
 
 Once we have the smallest program we can add it to the crash corpus in [`internal/vm/fuzz/`](../internal/vm/fuzz/) and running `make fuzz` should run and fail on it straight away.
 
-Or, variants of the program can be added to the various `*Invalid` tests in parts of the `vm` module, e.g. [`parser_test.go`](../internal/vm/parser/parser_test.go) or [`checker_test.go`](../internal/vm/checker/checker_test.go) depending on where in the compiler the defect is occuring.
+Or, variants of the program can be added to the various `*Invalid` tests in parts of the `vm` module, e.g. [`parser_test.go`](../internal/vm/parser/parser_test.go) or [`checker_test.go`](../internal/vm/compiler/checker/checker_test.go) depending on where in the compiler the defect is occuring.
 
 If the crash is in `vm.go` then we can dump the program to see what AST and types, and bytecode it generates.
