+++ internal/vm/codegen/codegen_test.go
@@ -10,10 +10,10 @@ import (
 	"time"
 
 	"github.com/google/mtail/internal/testutil"
-	"github.com/google/mtail/internal/vm/compiler/ast"
-	"github.com/google/mtail/internal/vm/checker"
 	"github.com/google/mtail/internal/vm/code"
 	"github.com/google/mtail/internal/vm/codegen"
+	"github.com/google/mtail/internal/vm/compiler/ast"
+	"github.com/google/mtail/internal/vm/compiler/checker"
 	"github.com/google/mtail/internal/vm/parser"
 )
