+++ internal/vm/compiler/checker/checker.go
